Add support for converting unique schemas to consistent format
