+++ tests/open_alchemy/table_args/test_factory.py
@@ -1,38 +1,116 @@
 """Tests for table args factories."""
 
-# import pytest
+import functools
 
-# from open_alchemy.table_args import factory
+import pytest
 
-
-# @pytest.mark.table_args
-# def test_handle_column_list():
-#     """
-#     GIVEN list of columns
-#     WHEN _handle_column_list is called
-#     THEN list of unique constraints is returned.
-#     """
-#     spec = ["column 1"]
-
-#     unique_list = factory._handle_column_list(  # pylint: disable=protected-access
-#         spec=spec
-#     )
-
-#     assert unique_list == [{"columns": ["column 1"]}]
+from open_alchemy import exceptions
+from open_alchemy.table_args import factory
 
 
-# @pytest.mark.parametrize(
-#     "spec",
-#     [[["column 1"]], [["column 1"], ["column 2"]]],
-#     ids=["single", "multiple"],
-# )
-# @pytest.mark.table_args
-# def test_handle_column_list_list(spec):
-#     """
-#     GIVEN list of list of columns
-#     WHEN _handle_column_list_list is called
-#     THEN indexes with columns are returned.
-#     """
-#     indexes = factory._handle_column_list_list(spec=spec)
+@pytest.mark.parametrize(
+    "spec, expected_name",
+    [
+        (["column 1"], "ColumnList"),
+        ([["column 1"]], "ColumnListList"),
+        ({"columns": ["column 1"]}, "UniqueConstraint"),
+        ({"columns": ["column 1"], "name": "name 1"}, "UniqueConstraint"),
+        ([{"columns": ["column 1"]}], "UniqueConstraintList"),
+        ({"name": "name 1", "expressions": ["column 1"]}, "Index"),
+        ({"name": "name 1", "expressions": ["column 1"], "unique": True}, "Index"),
+        ([{"name": "name 1", "expressions": ["column 1"]}], "IndexList"),
+    ],
+    ids=[
+        "ColumnList",
+        "ColumnListList",
+        "UniqueConstraint no name",
+        "UniqueConstraint",
+        "UniqueConstraintList",
+        "Index no unique",
+        "Index",
+        "IndexList",
+    ],
+)
+@pytest.mark.table_args
+def test_spec_to_schema_name(spec, expected_name):
+    """
+    GIVEN spec and expected name
+    WHEN _spec_to_schema_name is called with the spec
+    THEN the expected name is returned
+    """
+    name = factory._spec_to_schema_name(spec=spec)  # pylint: disable=protected-access
 
-#     mapped_indexes = list(map(lambda index: index._pending_colargs))
+    assert name == expected_name
+
+
+@pytest.mark.parametrize(
+    "schema_names, spec, raises",
+    [
+        ([], ["column 1"], True),
+        (["ColumnListList"], ["column 1"], True),
+        (["ColumnList"], ["column 1"], False),
+        (["ColumnListList", "UniqueConstraint"], ["column 1"], True),
+        (["ColumnList", "UniqueConstraint"], ["column 1"], False),
+        (["UniqueConstraint", "ColumnList"], ["column 1"], False),
+    ],
+    ids=[
+        "empty schemas,   -,              raises",
+        "single schemas,   spec diff,     raises",
+        "single schemas,   spec match,    not raises",
+        "multiple schemas, spec not in,   raises",
+        "multiple schemas, spec in,       not raises",
+        "multiple schemas, other spec in, not raises",
+    ],
+)
+@pytest.mark.table_args
+def test_spec_to_name_schemas(schema_names, spec, raises):
+    """
+    GIVEN list of schema names, spec and whether a raise is expected
+    WHEN _spec_to_schema_name is called with the spec and schemas
+    THEN SchemaNotFoundError is raised is a raise is expected.
+    """
+    test_func = functools.partial(
+        factory._spec_to_schema_name,  # pylint: disable=protected-access
+        spec=spec,
+        schema_names=schema_names,
+    )
+
+    if raises:
+        with pytest.raises(exceptions.SchemaNotFoundError):
+            test_func()
+    else:
+        test_func()
+
+
+@pytest.mark.parametrize(
+    "spec, expected_spec",
+    [
+        (["column 1"], [{"columns": ["column 1"]}]),
+        ([["column 1"]], [{"columns": ["column 1"]}]),
+        (
+            [["column 1"], ["column 2"]],
+            [{"columns": ["column 1"]}, {"columns": ["column 2"]}],
+        ),
+        ({"columns": ["column 1"]}, [{"columns": ["column 1"]}]),
+        ([{"columns": ["column 1"]}], [{"columns": ["column 1"]}]),
+    ],
+    ids=[
+        "column list",
+        "column list of list single",
+        "column list of list multiple",
+        "unique constraint",
+        "list of unique constraint",
+    ],
+)
+@pytest.mark.table_args
+def test_handle_unique(spec, expected_spec):
+    """
+    GIVEN specification and expected specification
+    WHEN _handle_unique is called with the specification
+    THEN the expected specification is returned.
+    """
+    returned_spec = factory._handle_unique(  # pylint: disable=protected-access
+        spec=spec
+    )
+
+    assert returned_spec == expected_spec
